Available translations are now sorted by locale. Minor changes for contributing translations. When importing a locale, empty translations are not added to DB, untranslated messages are fetched from l10n_messages. Ajax requests now show errors if any and, when searching for catalogs on CVS if only found an entry, keep searching until a file is found. Reduces search requests.

+++ l10nman/db_default.py
@@ -36,11 +36,11 @@ tables = [
     Table('l10n_locale_stats', key='locale_id')[
         Column('locale_id', type='integer'), # FK l10n_locales.id
         Column('translated', type='integer'),
-        Column('translated_percent', type='integer'),
+        Column('translated_percent', type='float'),
         Column('fuzzy', type='integer'),
-        Column('fuzzy_percent', type='integer'),
+        Column('fuzzy_percent', type='float'),
         Column('untranslated', type="integer"),
-        Column('untranslated_percent', type="integer")
+        Column('untranslated_percent', type="float")
     ],
     Table('l10n_messages', key='id')[
         Column('id', auto_increment=True),
